refactor majeur création de packages mesh, gameobject, objimporter et shaders
