+++ A3D/tp1/app/src/main/java/fr/univ_poitiers/dptinfo/algo3d/gameobject/Room.java
@@ -1,5 +1,7 @@
-package fr.univ_poitiers.dptinfo.algo3d;
+package fr.univ_poitiers.dptinfo.algo3d.gameobject;
 
+
+import fr.univ_poitiers.dptinfo.algo3d.mesh.Plane;
 
 public class Room extends GameObject {
