+++ A3D/tp1/app/src/main/java/fr/univ_poitiers/dptinfo/algo3d/gameobject/Transform.java
@@ -1,6 +1,8 @@
-package fr.univ_poitiers.dptinfo.algo3d;
+package fr.univ_poitiers.dptinfo.algo3d.gameobject;
 
 import android.opengl.Matrix;
+
+import fr.univ_poitiers.dptinfo.algo3d.Vec3f;
 
 public class Transform {
