+++ A3D/tp1/app/src/main/java/fr/univ_poitiers/dptinfo/algo3d/mesh/Pipe.java
@@ -1,12 +1,12 @@
-package fr.univ_poitiers.dptinfo.algo3d;
+package fr.univ_poitiers.dptinfo.algo3d.mesh;
 
-public class Cylinder extends Mesh {
-    public Cylinder(int quarter) {
-        vertexpos = new float[((2*2) * (quarter + 1)+2) * 3];
-        triangles = new int[quarter * 4 * 3];
+public class Pipe extends Mesh {
+    public Pipe(int quarter) {
+
+        vertexpos = new float[((2*2) * (quarter + 1)) * 3];
+        triangles = new int[quarter * 2 * 3 * 2];
         int k = 0;
         float r = 1.f;
-
         for (int i = 0; i <= quarter; i++) {
             double theta = Math.toRadians((360.0 / quarter) * i);
             float x = (float) (r * Math.cos(theta));
@@ -29,27 +29,43 @@ public class Cylinder extends Mesh {
             vertexpos[k++] = 1.f;
             vertexpos[k++] = z;
         }
-        vertexpos[vertexpos.length - 5] = 0.f;
-        vertexpos[vertexpos.length - 2] = 1.f;
 
         k = 0;
+        //faces internes
         for (int i = 0; i < quarter; i++) {
+            triangles[k++] =  (i * 2 + 1);
+            triangles[k++] =  (i * 2 + 2);
+            triangles[k++] =  (i * 2 + 3);
+            triangles[k++] =  (i * 2 + 1);
+            triangles[k++] =  (i * 2);
+            triangles[k++] =  (i * 2 + 2);
+        }
+
+        //faces externes
+        for (int i = quarter+1; i < quarter*2+1; i++) {
             triangles[k++] =  (i * 2 + 1);
             triangles[k++] =  (i * 2 + 3);
             triangles[k++] =  (i * 2 + 2);
             triangles[k++] =  (i * 2 + 1);
             triangles[k++] =  (i * 2 + 2);
             triangles[k++] =  (i * 2);
-            triangles[k++] =  ((i+quarter+1)* 2 + 1);
-            triangles[k++] =  (vertexpos.length/3-1);
-            triangles[k++] =  ((i+quarter+1) * 2 + 3);
-            triangles[k++] =  ((i+quarter+1) * 2);
-            triangles[k++] =  ((i+quarter+1) * 2 + 2);
-            triangles[k++] =  (vertexpos.length/3-2);
         }
 
         k=0;
-        normals = new float[vertexpos.length];
+
+        normals = new float[((2) * (quarter + 1)) * 3*2];
+
+        for (int i = 0; i <= quarter; i++) {
+            double theta = Math.toRadians((360.0 / quarter) * i);
+            float x = (float) (r * Math.cos(theta));
+            float z = (float) (r * Math.sin(theta));
+            normals[k++] = -x;
+            normals[k++] = 0.f;
+            normals[k++] = -z;
+            normals[k++] = -x;
+            normals[k++] = 0.f;
+            normals[k++] = -z;
+        }
 
         for (int i = 0; i <= quarter; i++) {
             double theta = Math.toRadians((360.0 / quarter) * i);
@@ -62,15 +78,7 @@ public class Cylinder extends Mesh {
             normals[k++] = 0.f;
             normals[k++] = z;
         }
-        for (int i = 0; i <= quarter; i++) {
-            normals[k++] = 0.f;
-            normals[k++] = -1.f;
-            normals[k++] = 0.f;
-            normals[k++] = 0.f;
-            normals[k++] = 1.f;
-            normals[k++] = 0.f;
-        }
-        normals[normals.length - 5] = -1.f;
-        normals[normals.length - 2] = 1.f;
+
+
     }
 }
