+++ A3D/tp1/app/src/main/java/fr/univ_poitiers/dptinfo/algo3d/objimporter/OBJFace.java
@@ -1,4 +1,4 @@
-package fr.univ_poitiers.dptinfo.algo3d;
+package fr.univ_poitiers.dptinfo.algo3d.objimporter;
 
 public class OBJFace {
