+++ A3D/tp1/app/src/main/java/fr/univ_poitiers/dptinfo/algo3d/shaders/PixelShaders.java
@@ -1,9 +1,8 @@
-package fr.univ_poitiers.dptinfo.algo3d;
+package fr.univ_poitiers.dptinfo.algo3d.shaders;
 
 import android.content.Context;
 
-public class BlinnPhongShaders extends LightingShaders{
-
+public class PixelShaders extends LightingShaders{
 
     /**
      * Fragment shader
@@ -14,9 +13,6 @@ public class BlinnPhongShaders extends LightingShaders{
                     +"uniform mat4 uProjectionMatrix;\n"
                     +"uniform mat3 uNormalMatrix;\n"
                     // Light source definition
-                    +"uniform float uConstantAttenuation;\n"
-                    +"uniform float uLinearAttenuation;\n"
-                    +"uniform float uQuadraticAttenuation;\n"
                     +"uniform vec4 uAmbiantLight;\n"
                     +"uniform bool uLighting;\n"
                     +"uniform vec3 uLightPos;\n"
@@ -24,10 +20,6 @@ public class BlinnPhongShaders extends LightingShaders{
                     // Material definition
                     +"uniform bool uNormalizing;\n"
                     +"uniform vec4 uMaterialColor;\n"
-                    //Specular effect
-                    +"uniform float uMaterialShininess;\n"
-                    +"uniform vec4 uLightSpecular;\n"
-                    +"uniform vec4 uMaterialSpecular;\n"
 
                     +"varying vec4 posf;\n"
                     +"varying vec3 normalf;\n"
@@ -35,23 +27,10 @@ public class BlinnPhongShaders extends LightingShaders{
                     +"void main(void) {\n"
                     +"  if (uLighting)\n"
                     +"  {\n"
-                    +"    float distance = length(uLightPos-posf.xyz);\n"
-                    +"    float attenuation = 1.0 /(uConstantAttenuation + uLinearAttenuation* distance + uQuadraticAttenuation * (distance * distance)) ;\n"
-                    +"    vec3 viewdir=normalize(-posf.xyz);\n"
                     +"    vec3 normal = normalize(normalf);"
                     +"    vec3 lightdir=normalize(uLightPos-posf.xyz);\n"
-
-                    +"    vec3 halfdir = normalize(lightdir + viewdir);\n"
-
                     +"    float weight = max(dot(normal, lightdir),0.0);\n"
-                    +"    vec4 dColor = uMaterialColor*(uAmbiantLight+weight*uLightColor);\n"
-
-                    +"    float spec = pow(max(dot(halfdir, normal), 0.0), uMaterialShininess*4.0);\n"
-                    +"    vec4 specColor = uMaterialSpecular*uLightSpecular*spec;\n"
-
-                    +"    dColor *= attenuation;\n"
-                    +"    specColor *= attenuation;\n"
-                    +"    gl_FragColor = dColor + specColor;\n"
+                    +"    gl_FragColor = uMaterialColor*(uAmbiantLight+weight*uLightColor);\n"
                     +"  }\n"
                     +"  else gl_FragColor = uMaterialColor;\n"
                     +"}\n";
@@ -88,18 +67,20 @@ public class BlinnPhongShaders extends LightingShaders{
                     +"  gl_Position= uProjectionMatrix*posf;\n"
                     +"}\n";
 
+
+
     /**
      * Constructor. nothing to do, everything is done in the super class...
      *
      * @param context
      */
-    public BlinnPhongShaders(Context context) {
+    public PixelShaders(Context context) {
         super(context);
     }
 
+
     @Override
     public int createProgram(Context context) {
-        return initializeShadersFromResources(context,"blinn_phong_vert.glsl","blinn_phong_frag.glsl");
+        return initializeShadersFromResources(context,"pixel_vert.glsl","pixel_frag.glsl");
     }
-
 }
