+++ A3D/tp1/app/src/main/java/fr/univ_poitiers/dptinfo/algo3d/shaders/ShadingMode.java
@@ -1,4 +1,4 @@
-package fr.univ_poitiers.dptinfo.algo3d;
+package fr.univ_poitiers.dptinfo.algo3d.shaders;
 
 public enum ShadingMode {
     FLAT_SHADING,
